Add alternate light entry
